Shoot from max focus distance to min focus distance
Before this commit you first set minimum focus. After that you set maximum focus. Therefore the focus
ring is in the maximum focus position. When starting shooting the app
first moves focus to the minimum position. On 90mm lens this takes a long
time. If we start shooting from the maximum position and move to the
minimum after we will save some time at the beginning.

This commit reverses the shooting order so images are captured from the
maximum focus distance to the minimum focus distance and therefore
skipping the part where focus is being rolled all over the focus range at
the beginning.